add confirm on delete

+++ src/components/task.jsx
@@ -55,7 +55,10 @@ export default function Task({task}) {
 
     const onDeleteClick = e => {
         // TODO add confirmation
-        deleteTask(task.id)
+        const isConfirm = window.confirm('Supprimer la tâche ?');
+        if (isConfirm) {
+            deleteTask(task.id)
+        }
         e.stopPropagation();
     }
